[DRONE] Use cbindgen to generate C header for PRU
PRUs and drone rust code use the shared memory for sharing data.
The data structure was defined in both C and Rust code, easily leading
to mistakes if one definition was modified alone.

cbindgen generates a C header for the PRU so that there is only one
definition for the shared data structure.