[MODEL] Move PID computation out of ODE
Previous implementation was computing the PID inside the ODE and was
difficult to extend. New implementation compute the PID outside of the
ODE and call the ODE solver to compute the drone model between each PID
step. Additionaly this commit introduce the possibility to run with a
custom thrust command as a numpy array.